19-IMPROVEMENT Move getDueDate to classroom-api. Remove unused call.
