+++ classroom-api/src/index.ts
@@ -1,4 +1,4 @@
-import {ClassroomApi} from './classroom-api'
+import {ClassroomApi, getDueDate} from './classroom-api'
 import {StudentList} from './students'
 import {Authenticator} from './authenticate'
 import {Submission} from 'dt-types'
@@ -13,6 +13,7 @@ export * from './mailer'
 export { downloadAll, downloadSome, downloadAtInterval } from './downloadHW'
 export { Authenticator } from './authenticate'
 export { downloadError } from './classroom-api'
+export { getDueDate } from './classroom-api'
 
 export async function getSubmissions(subject: string, homework: string, studentList: StudentList, auth: Authenticator): Promise<Submission[]> {
 	let classroom = await ClassroomApi.findClass(subject, auth)
@@ -30,15 +31,4 @@ export async function getSubmissions(subject: string, homework: string, studentL
 	}))
 
 	return submissions.filter(response => studentList.getStudentById(response.userId!)).map(s => fromResponse(s, studentList))
-}
-
-export async function getDueDate(subject: string, homeworkTitle: string, auth: Authenticator): Promise<Date> {
-	let classroom = await ClassroomApi.findClass(subject, auth);
-	const courseWork = await classroom.listCourseWork();
-
-	return courseWork.filter(work => work.title === homeworkTitle).map((work): Date =>{
-		if(work.dueDate === undefined || work.dueDate.year === undefined  || work.dueDate.month === undefined || work.dueDate.day === undefined )
-			throw "Selected homework does not have due date"
-		return new Date(work.dueDate.year, work.dueDate.month, work.dueDate.day)
-	})[0]
 }
